requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
wikipedia
langchain-text-splitters
langchain-openai
chromadb
ollama
sentence-transformers
langchain-huggingface
faiss-cpu
langchain-ollama
langchain-chroma
beautifulsoup4
streamlit
langchain-groq
langchain-core
fastapi
uvicorn
langserve